variable `CTA_H` & fix qkv bias

+++ src/turbomind/kernels/attention/kv_cache_utils_v2.cu
@@ -22,10 +22,10 @@ __global__ void __launch_bounds__(128) ProcessKV_v2(char**       blocks,
                                                     const int*   cu_block_num,
                                                     const float* rope_base,
                                                     float        rope_ti_scale,
-                                                    int          stride_b,
-                                                    int          stride_c,
-                                                    int          stride_h,
-                                                    int          stride_s,
+                                                    int64_t      stride_b,
+                                                    int64_t      stride_c,
+                                                    int64_t      stride_h,
+                                                    int64_t      stride_s,
                                                     int          layer_id,
                                                     BlockLayout  block_layout)
 {
@@ -83,9 +83,9 @@ __global__ void __launch_bounds__(128) ProcessKV_v2(char**       blocks,
     for (int s = 0; s < ITER_S; ++s) {
         PRAGMA_UNROLL
         for (int c = 0; c < ITER_C; ++c) {
-            const int qi = offset.y + s * Map::kDeltaS + token_idx;  // sequence local
-            const int di = offset.x + c * Map::kDeltaC;
-            const int index =
+            const int     qi = offset.y + s * Map::kDeltaS + token_idx;  // sequence local
+            const int     di = offset.x + c * Map::kDeltaC;
+            const int64_t index =
                 (batch_idx * stride_b + qi_beg * stride_c + qi * stride_s + head_idx * stride_h) * HeadDim + di;
             if (qi < q_len) {
                 Ldg(vec_K[s][c], &k[index]);
@@ -196,10 +196,10 @@ void invokeProcessKV_v2(char**       blocks,
                         const int*   cu_block_num,
                         const float* rope_base,
                         float        rope_ti_scale,
-                        int          stride_b,
-                        int          stride_c,
-                        int          stride_h,
-                        int          stride_s,
+                        int64_t      stride_b,
+                        int64_t      stride_c,
+                        int64_t      stride_h,
+                        int64_t      stride_s,
                         int          block_seq_len,
                         int          layer_id,
                         int          max_q_len,
@@ -263,10 +263,10 @@ void invokeProcessKV_v2(char**       blocks,
                                      const int*   cu_block_num,                                                        \
                                      const float* rope_base,                                                           \
                                      float        rope_ti_scale,                                                       \
-                                     int          stride_b,                                                            \
-                                     int          stride_c,                                                            \
-                                     int          stride_h,                                                            \
-                                     int          stride_s,                                                            \
+                                     int64_t      stride_b,                                                            \
+                                     int64_t      stride_c,                                                            \
+                                     int64_t      stride_h,                                                            \
+                                     int64_t      stride_s,                                                            \
                                      int          block_seq_len,                                                       \
                                      int          layer_id,                                                            \
                                      int          max_q_len,                                                           \
@@ -289,10 +289,10 @@ __global__ void __launch_bounds__(128) flattenKV_v2(T*           k,
                                                     const int*   cu_block_num,
                                                     const float* rope_base,
                                                     float        rope_ti_scale,
-                                                    int          stride_b,
-                                                    int          stride_c,
-                                                    int          stride_h,
-                                                    int          stride_s,
+                                                    int64_t      stride_b,
+                                                    int64_t      stride_c,
+                                                    int64_t      stride_h,
+                                                    int64_t      stride_s,
                                                     int          layer_id,
                                                     BlockLayout  block_layout)
 {
@@ -390,9 +390,9 @@ __global__ void __launch_bounds__(128) flattenKV_v2(T*           k,
     for (int s = 0; s < ITER_S; ++s) {
         PRAGMA_UNROLL
         for (int c = 0; c < ITER_C; ++c) {
-            const int si = offset.y + s * Map::kDeltaS + token_idx;
-            const int di = offset.x + c * Map::kDeltaC;
-            const int index =
+            const int     si = offset.y + s * Map::kDeltaS + token_idx;
+            const int     di = offset.x + c * Map::kDeltaC;
+            const int64_t index =
                 (batch_idx * stride_b + ti_beg * stride_c + si * stride_s + head_idx * stride_h) * HeadDim + di;
             if (si < seq_len) {
                 Store(&k[index], out_K[s][c]);
@@ -410,10 +410,10 @@ void invokeFlattenKV_v2(T*           k,
                         const int*   cu_block_num,
                         const float* rope_base,
                         float        rope_ti_scale,
-                        int          stride_b,
-                        int          stride_c,
-                        int          stride_h,
-                        int          stride_s,
+                        int64_t      stride_b,
+                        int64_t      stride_c,
+                        int64_t      stride_h,
+                        int64_t      stride_s,
                         int          block_seq_len,
                         int          layer_id,
                         int          max_seq_len,
@@ -471,10 +471,10 @@ void invokeFlattenKV_v2(T*           k,
                                      const int*   cu_block_num,                                                        \
                                      const float* rope_base,                                                           \
                                      float        rope_ti_scale,                                                       \
-                                     int          stride_b,                                                            \
-                                     int          stride_c,                                                            \
-                                     int          stride_h,                                                            \
-                                     int          stride_s,                                                            \
+                                     int64_t      stride_b,                                                            \
+                                     int64_t      stride_c,                                                            \
+                                     int64_t      stride_h,                                                            \
+                                     int64_t      stride_s,                                                            \
                                      int          block_seq_len,                                                       \
                                      int          layer_id,                                                            \
                                      int          max_seq_len,                                                         \
